% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_mananciais_sabesp.R
\name{pega_manaciais_sabesp}
\alias{pega_manaciais_sabesp}
\title{Pega os Dados Pluviometricos da Sabesp}
\usage{
pega_manaciais_sabesp(data_coleta = Sys.Date())
}
\arguments{
\item{data_coleta}{Data de coleta dos dados no formato "YYYY-MM-DD".}
}
\value{
dataset com os dados pluviometricos do Sabesp
}
\description{
Pega os Dados Pluviometricos da Sabesp
}
\examples{
pega_manaciais_sabesp(data_coleta = Sys.Date())

# pega os o ultimo mes
data_coleta = Sys.Date()-1:30

# pega para a ultima semana
# data_coleta = Sys.Date()-1:7
}
