% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_mananciais_sabesp.R
\name{pega_mananciais_sabesp}
\alias{pega_mananciais_sabesp}
\title{Pega os dados pluviométricos da sabesp}
\usage{
pega_mananciais_sabesp(data_coleta = Sys.Date())
}
\arguments{
\item{data_coleta}{data da coleta no forma YYYY-MM-DD}
}
\value{
dataset com dados da sabesp
}
\description{
Pega os dados pluviométricos da sabesp
}
\examples{
#pega_mananciais_sabesp(data_coleta = Sys.Date())

#pega para o ultimo mes
#pega_mananciais_sabesp(data_coleta = Sys.Date()-1:30
}
